archive/2025/haunted_porch/hauntedPorch_controller/requirements.txt
fastapi
uvicorn
paho-mqtt
